DRY up media sesh update
